requirements.txt
black>=23.1.0,<23.2
kafka-python
faust-streaming>=0.10.4,<0.11